fixed PALLADIO-233: Uniform distribution does not work properly
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Commons/trunk@17743 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ de.uka.ipd.sdq.probfunction.math/src/de/uka/ipd/sdq/probfunction/math/apache/distribution/UniformIntDistributionImpl.java
@@ -55,7 +55,7 @@ public class UniformIntDistributionImpl extends AbstractIntegerDistribution {
 		else if(x>b)
 			return 1;
 		
-		return (x-a+1)/intCount;
+		return (x-a+1)/(double)intCount;
 	}
 
 	/* (non-Javadoc)
